seshat: Change the type of the return value for add_backlog_events().
Events from the backlog are mostly added if a sync with a limited room
timeline happens. The client will need to fetch events from the room
history in that case.

To signal when the client catches up with the existing indexed
history we return a boolean to indicate if all the added messages were
already stored and indexed or not.