seshat: Add a Directory implementation that transparently encrypts files.
This patch adds implementation of a Tantivy Directory that wraps the
standard MmapDirectory from Tantivy and transparently adds authenticated
encryption on top of it.

All the reads and writes will go through an AES-HMAC based
encryptor/decryptor.

The crypto primitives were chosen to be similar to the primitives given
in the "Key exports" chapter[1] of the Matrix Client-Server API.

The new directory implementation isn't used anywhere yet but will
utilized in the Seshat index in a later patch.

[1] https://matrix.org/docs/spec/client_server/r0.5.0#key-exports