seshat: Add support to encrypt the database.
This patch adds SQLCipher support to our event database. SQLCipher
transparently encrypts the whole database including the table names.

Users are still able to poke the database manually with the sqlcipher
command line tool which acts as a drop in replacement to the sqlite
command line tool.

To encrypt the database a user should open the Seshat database with a
passphrase set in the config. If no database exists a new encrypted one
will be created. If one exists Seshat will try to open it with the given
passphrase.