seshat: Use the EncryptedMmapDirectory in the Seshat index.
This patch utilizes the EncryptedMmapDirectory in our Index. The user
provided API is the same as for the database, if a passphrase is
provided in the config the index will be encrypted.

The Seshat database is also expanded to encrypt the index it creates if
the database is as well encrypted.