seshat: Don't store the content_value in the database.
We store the whole event serialized, and let the caller decide how it
should be serialized. While this puts some responsibilty on the caller,
it allows great flexibility in the case of extensible or custom events.

Storing the value of the content separately ends up storing almost the
whole event twice in the database, the serialized one and the differing
event key/values that are store independently.

The value of the content isn't used anywhere internally so we can safely
remove it from the database. While we do load Event objects from the
database, they are only used internally so it is safe to initialize them
with an empty content value.