index: Don't allow users to commit events to the index too often.
Tantivy doesn't behave nicely if commit() is called too often on the index
writer. A commit means that Tantivy will spawn threads that will try to merge
index segments together, that is, it tries to merge a bunch smaller files into
one larger.

If users call commit faster than those threads manage to merge the segments,
the number of spawned threads keeps on increasing.

This patch limits the number of commit calls we forward to the index writer.
To avoid data loss we mark the uncommitted events as such in the database and
restore them if needed after a new database object is created.

Fixes: #21.