database: Make the resume_committing test non-flaky.
Teh resume_comitting test closes and opens the database immediately one
after another. Tantivy might still be in the process of being shut down
and hold on to the write lock.

Opening the database will fail in this case and the test will fail as
well.

This patch retries opening the database a couple of times with a slight
delay between the retries.