database: Add the content value back to the database.
This is a revert of a0119351b163c6cabdc7257fa8ab829c4e2f8393.

While we still do not use the content value internally, it will be used
once we need to rebuild our index.

Rebuilding the index means deleting all documents from Tantivy and
re-adding them, for this to work we would need to have the content
value.

We could de-serialize the content from the serialized event but this
would put the burden of serialization and de-serialization onto Seshat.

This would in turn pull in Serde as a dependency and make extensible
events hard to support.

Dropping the serialized representation of the event from the database
would make extensible events a no-go as well.