index: Don't forward commits to Tantivy if there's nothing to do.
Commits trigger a rotation of the writer threads and also triggers a
reload of the reader.

While the first case is harmless it still wastes CPU cycles
unnecessarily.

The second case currently behaves in a leaky manner unless a searcher
is fetched from Tantivy. While this patch gets rid of the leaky manner
if there's nothing to do, unhindered memory growth will occur if events
are added.