index: Limit the number of writer threads.
A commit rotates the writer threads, those threads separately may write
files. A reload in the OnCommit mode happens automatically if the list
of files changes in the directory.

To slow down the leaky behaviour we can limit the number of threads to
also limit the number of files that get written to the indexer
directory.