seshat: Enable WAL mode for the database.
In the default journaling mode SQLite can end up throwing SQLITE_BUSY
errors in the case of parallel access.

We have a setup with a single writer thread and possibly multiple reader
threads. This is the perfect setup for WAL mode and as per docs[1] will
remove the chance of getting SQLITE_BUSY errors.

[1] https://www.sqlite.org/wal.html