seshat: Don't share a connection between threads without a mutex.
The searcher database connection is reused to avoid opening and
potentially failing to open a new connection. This means that multiple
searches will reuse the same connection and possibly in different
threads.

This can lead to undefined behaviour, it was mostly fine since
clients will do one search at a time.