index: Don't fill out the whole 128 bit counter with randomness.
This patch changes the way our nonces/IVs are created, while it is quite
unlikely that we would encounter an overflow of our counter, considering the
files Tantivy creates, this makes sure that it can't happen.

The upper 64 bits are populated by randomness while the lower parts are
left 0 initialized.

This aligns the set up to how the counter is created in the attachment
encryption part of the Matrix spec as well.