database: Call the shut down method when we change the passphrase
Changing the passphrase involves a close/open cycle of the database.
This means if we don't shut down the database properly we'll end up
with a race where we might try to open a locked database.

The `shutdown()` method already closes all database connections and
waits for all threads to be terminated, so use the method after the
passphrase has been changed.