Allow SQLCipher to upgrade it's cipher settings
SQLCipher version 4.0[1] changed the default cipher
settings, which means that after an upgrade, users will fail to open a
database which was created using previous SQLCipher versions.

The cipher_migrate PRAGMA can be used to migrate the cipher settings
to the new default values. This allows users to transparently continue
on using their existing database.

This patch implements the algorithm described in the cipher_migrate
documentation[1].

[1]: https://github.com/sqlcipher/sqlcipher/blob/master/CHANGELOG.md#400---november-2018---400-changes
[2]: https://www.zetetic.net/sqlcipher/sqlcipher-api/#cipher_migrate